refactor: Product 클래스의 price 자료형 변경

+++ src/main/java/dev/nbcsparta/assignment/commerce_backoffice/dto/CreateProductRequest.java
@@ -13,7 +13,7 @@ public record CreateProductRequest(
         String category,
 
         @NotNull
-        Long price,
+        int price,
 
         @NotNull
         int quantity,
